+++ index.js
@@ -11,10 +11,13 @@
   */
 "use strict";
 
-const [{createServer}, {fork}, {resolve}, cluster] = [require("net"), require("child_process"), require("path"), require("cluster")];
+const [{createServer}, {fork}, {resolve}, cluster, co] = [require("net"), require("child_process"), require("path"), require("cluster"), require("co")];
 const {get} = require(resolve(__dirname, "config"));
 
 const [threadManager, slavePath ,bindHost, registryPort, {enable, slaveNum = 1}, remote] = [new Map(), resolve(__dirname, "distributed","slave"), get("bindHost"), get("registryPort"), get("distributed"), get("remote")];
+
+const NODELOCK = Symbol("lock");
+module[NODELOCK] = {};
 
 const startSlave = socket => {
   const worker = fork(slavePath);
@@ -54,6 +57,45 @@ if (enable) {
   console.log("portal server engine use sing thread module");
   require(resolve(__dirname, "distributed", "single"));
 }
+
+process.on("hasLock", (lock, callback) => {
+  const flag = module[NODELOCK][lock];
+  if (callback && callback.apply) {
+    callback(flag);
+  }
+});
+
+process.on("tryObmitLock", (lock, genertaor, callback) => {
+  module[NODELOCK][lock].push(genertaor);
+  if (callback) {
+    process.emit("unLock", lock);
+  }
+});
+
+process.on("lock", (lock, callback) => {
+  module[NODELOCK][lock] = [];
+  if (callback && callback.apply) {
+    callback();
+  }
+});
+
+process.on("unLock", lock => {
+  const tasks = module[NODELOCK][lock];
+  if (tasks && tasks.length > 0) {
+    let funs = tasks.shift();
+    co(function*() {
+      const value = yield funs;
+      return value;
+    }).then(() => {
+      process.emit("unLock", lock);
+    }).catch(err => {
+      console.log(err);
+      process.emit("unLock", lock);
+    });
+  } else {
+    delete module[NODELOCK][lock];
+  }
+});
 
 process.on("exit", () => {
   for (let worker of threadManager.values()) {